Added basic Status handlers
Refactored the loading & error reducers, adding a new "status" one and removing the loadingReducer altogether.
The main idea with this new structure is that we keep track of every status of an action: `NOT_STARTED`, `LOADING`, `SUCCESS` and `ERROR`.
Then, using selectors, we can get the current status of an action at any point of it's flow.
This should be usefull when reacting to successfull submits, CRUDs, etc.